feat: implement Insights component and integrate into Settings
