+++ src/components/Insights.tsx
@@ -1,11 +1,40 @@
-import { useState } from "react";
-import { useSettingsContext } from "../hooks/useSettingsContext";
 import { Trash, X } from "@phosphor-icons/react";
+import { useSettingsContext } from "../hooks/useSettingsContext";
+import Tabs from "./Tabs";
+import { useState } from "react";
 import classNames from "classnames";
-import { generateDummyHistory } from "../utility/dummyData"
+import { generateDummyHistory } from "../utility/dummyData";
+import { eachDayOfInterval, endOfWeek, format, isWithinInterval, startOfWeek } from "date-fns";
+import { Bar, BarChart, ResponsiveContainer, Tooltip, XAxis } from "recharts";
 
-export default function History() {
-  const { setShowHistory, history, setHistory, } = useSettingsContext();
+export default function Insights() {
+  const { setShowInsights } = useSettingsContext();
+
+  return (
+    <div className="w-full">
+      <header className="flex justify-between items-center">
+        <h2 className="text-2xl font-semibold">Insights</h2>
+
+        <button
+          type="button"
+          className="bg-slate-200 p-1.5 rounded-full hover:bg-slate-300 inline-flex items-center gap-2 justify-center text-center aspect-square transition-colors font-semibold"
+          onClick={() => setShowInsights(false)}
+        >
+          <X weight="bold" color="currentColor" size={16} />
+        </button>
+      </header>
+
+      <Tabs tabs={[
+        { label: "Analytics", content: <Analytics /> },
+        { label: "History", content: <History /> },
+      ]}
+      />
+    </div>
+  )
+}
+
+function History() {
+  const { setShowInsights, history, setHistory, } = useSettingsContext();
   const [alert, setAlert] = useState(false);
 
   const downloadHistoryCSV = () => {
@@ -94,7 +123,7 @@ export default function History() {
         </div>
 
         {history.map((entry, index) => (
-          <div key={`${entry.date}-${index}`} role="row" className="relative grid py-2 grid-cols-5 gap-8">
+          <div key={`${entry.date}-${index}`} role="row" className="relative grid py-2.5 px-3 grid-cols-5 gap-8">
             <div role="gridcell" className="w-max">
               {new Date(entry.date).toLocaleDateString("de-DE", { weekday: "short", year: "numeric", month: "2-digit", day: "2-digit" })}
             </div>
@@ -118,21 +147,9 @@ export default function History() {
 
   return (
     <div className="w-full">
-      <header className="flex justify-between items-center">
-        <h2 className="text-2xl font-semibold">History</h2>
-
-        <button
-          type="button"
-          className="bg-slate-200 p-1.5 rounded-full hover:bg-slate-300 inline-flex items-center gap-2 justify-center text-center aspect-square transition-colors font-semibold"
-          onClick={() => setShowHistory(false)}
-        >
-          <X weight="bold" color="currentColor" size={16} />
-        </button>
-      </header>
-
-      <main className="flex flex-col gap-4 my-8 max-h-96 overflow-scroll">
+      <main className="flex flex-col gap-4 my-8 max-h-80 overflow-scroll border border-slate-200 rounded-lg">
         <div role="table" aria-label="History entries" className="divide-y divide-slate-200">
-          <div role="row" className="grid grid-cols-5 gap-8 font-semibold py-1">
+          <div role="row" className="grid grid-cols-5 gap-8 font-semibold py-1.5 sticky top-0 bg-slate-200 z-10 px-3 text-sm">
             <span role="columnheader">Date</span>
             <span role="columnheader">Starting at</span>
             <span role="columnheader">Ending at</span>
@@ -157,7 +174,7 @@ export default function History() {
           onClick={() => {
             if (alert) {
               setHistory([]);
-              setShowHistory(false);
+              setShowInsights(false);
             } else {
               setAlert(true);
             }
@@ -177,5 +194,95 @@ export default function History() {
         )}
       </footer>
     </div>
+  );
+}
+
+function Analytics() {
+  const { history } = useSettingsContext();
+
+  const currentDate = new Date();
+  const weekStart = startOfWeek(currentDate, { weekStartsOn: 1 }); // 1 represents Monday
+  const weekEnd = endOfWeek(currentDate, { weekStartsOn: 1 });
+
+  // Calculate total hours worked this week
+  const weeklyTotal = history
+    .filter((entry) =>
+      isWithinInterval(new Date(entry.date), { start: weekStart, end: weekEnd })
+    )
+    .reduce((total, entry) => {
+      return total + (entry.endTime - entry.startTime) / (1000 * 60 * 60);
+    }, 0);
+
+  // Calculate average session duration
+  const averageSessionDuration =
+    history.length > 0
+      ? history.reduce((total, entry) => {
+        return total + (entry.endTime - entry.startTime);
+      }, 0) / (history.length * 1000 * 60 * 60)
+      : 0;
+
+  // Prepare data for daily hours chart
+  const dailyData = eachDayOfInterval({ start: weekStart, end: weekEnd }).map(
+    (day) => {
+      const dayTotal = history
+        .filter(
+          (entry) =>
+            format(new Date(entry.date), "yyyy-MM-dd") ===
+            format(day, "yyyy-MM-dd")
+        )
+        .reduce((total, entry) => {
+          return total + (entry.endTime - entry.startTime) / (1000 * 60 * 60);
+        }, 0);
+
+      return {
+        name: format(day, "EEE"), // Day name (Mon, Tue, etc.)
+        hours: Number(dayTotal.toFixed(2)),
+      };
+    }
+  );
+
+  return (
+    <div className="w-full max-w-4xl">
+      <main className="flex flex-col gap-8 mt-8">
+        <div className="grid grid-cols-2 gap-8">
+          <div>
+            <label className="font-semibold mb-2">Weekly Total</label>
+            <h3 className="text-3xl font-bold">{weeklyTotal.toFixed(1)} h</h3>
+          </div>
+
+          <div>
+            <label className="font-semibold mb-2">Avg. Session Duration</label>
+            <h3 className="text-3xl font-bold">
+              {averageSessionDuration.toFixed(1)} h
+            </h3>
+          </div>
+        </div >
+
+        <div>
+          <label className="font-semibold mb-4"> Daily Hours Worked</label>
+          <div className="h-96 max-h-96">
+            <ResponsiveContainer width="100%" height="100%" className="">
+              <BarChart data={dailyData}>
+                <XAxis dataKey="name" />
+                <Tooltip
+                  wrapperClassName="rounded-md shadow-lg"
+                  labelClassName="var(--color-slate-700) font-semibold"
+                  contentStyle={{
+                    background: "var(--color-slate-100)",
+                    color: "var(--color-slate-700)",
+                    borderColor: "transparent",
+                    padding: "0.5rem",
+                    lineHeight: "1",
+                    margin: "0"
+                  }}
+                  cursor={{ fill: "var(--color-slate-300)" }}
+                />
+                <Bar dataKey="hours" fill="var(--color-slate-700)" />
+              </BarChart>
+            </ResponsiveContainer>
+          </div>
+        </div >
+      </main >
+    </div >
   );
 }
