+++ src/contexts/SettingsContextProvider.tsx
@@ -15,8 +15,7 @@ export default function SettingsContextProvider({
 		useState<NotificationPermission>("default")
 
 	const [showSettings, setShowSettings] = useState(false)
-	const [showHistory, setShowHistory] = useState(false)
-	const [showAnalytics, setShowAnalytics] = useState(false)
+	const [showInsights, setShowInsights] = useState(false)
 
 	const [timeState, setTimeState] = useState<TimeState>(() => {
 		const saved = localStorage.getItem("timeState")
@@ -117,10 +116,8 @@ export default function SettingsContextProvider({
 			value={{
 				showSettings,
 				setShowSettings,
-				showHistory,
-				setShowHistory,
-				showAnalytics,
-				setShowAnalytics,
+				showInsights,
+				setShowInsights,
 				notificationPermission,
 				setNotificationPermission,
 				timeState,
